fixed issue with process using file

+++ Lab_RateMyProfessor/File_Manager.cs
@@ -23,18 +23,16 @@ namespace Lab_RateMyProfessor
             } else
             {
                 string path = "categories.json";
-                using (StreamWriter w = File.AppendText("categories.json"))
+                //using (StreamWriter w = File.AppendText("categories.json"))
+                if (!File.Exists(path))
+                {
+                    // Create a file to write to.
+                    string createText = "" + Environment.NewLine;
+                    File.WriteAllText(path, createText);
+                }
 
-                    /*
-                     * if (!File.Exists(path))
-        {
-            // Create a file to write to.
-            string createText = "Hello and Welcome" + Environment.NewLine;
-            File.WriteAllText(path, createText);
-        }
-                    */
 
-                try
+                    try
                 {
                     string existingJson = File.ReadAllText(path);
 
@@ -73,7 +71,7 @@ namespace Lab_RateMyProfessor
         public static bool addRating(Ratings rat)
         {
             string path = "ratings.json";
-            using (StreamWriter w = File.AppendText("ratings.json"))
+            //using (StreamWriter w = File.AppendText("ratings.json"))
 
             try
             {
@@ -118,7 +116,7 @@ namespace Lab_RateMyProfessor
             } else
             {
                 string path = "professors.json";
-                using (StreamWriter w = File.AppendText("professors.json"))
+                //using (StreamWriter w = File.AppendText("professors.json"))
 
                 try
                 {
